fix(ui): renderer loading deferred into own fn
